Fixed operator names in cap_tee files
The operator's function names generated as part of asm files/dump files is not correct. For eg: camlLwt_result__>|= is replaced by camlLwt_result__.3e.7c.3d_1789. So the compartment IDs are not updated properly leading to error. For now doing this in manual way. Possible solution is to generate cap_tee_files based on compiler output instead of generating them based on the generated asm files